Auto fill if one suggestion, Add user defined links to bottom of menu
This branch is for use on a personal project so not commited to
mastered. It adds two (optional) arguments to suggestiveInput;
required, and function Links. If required is set to "1", when user is
typing, if there is one suggestion and they blur the field, it will
auto fill the last suggestion.  FunctionLinks is an associative array
[Link Name : javascript] where every Link name is added to the bottom
of the menu, and when pressed, javascript is executed.